feat: 抽离 TypeExtensions.cs
